Implement ports
This commit implements the `ports` part of the architecture, using the following structure:

input & out ports are interfaces where input ports implemented in the application hexagon by the user cases, and output ports implemented by the adapters hexagon.

The cucumber tests are currently added to the repository, but cucumber does not work, there is an issue to fix the test and make them work properly.