Implement adapters
This commit implements the adapters for the project. They're exposing all the main operations that will later on be supported by the application. However, in this commit, none of them is still wired to the actual application, and there is no main in the JAR file yet.

All of that will arrive in a later commit.

Moreover, this commit adds the `bootstrap` module which is responsible for... well... bootstrap the spring boot application :smile: