Add "e2e" tests
This commit adds tests that allow testing the flow all the way from the input adapter up to the output adapter. It allows checking the connective between the different components and the behaviour of the system.

However, at this point, the test does not leverage any spring features and relay on manual dependency injection.