Extract spring specific to `bootstrap` module
This commit extracts away everything that is springboot-specific from the adapters and application, so the actual application is completely agnostic to the web framework and can be easily switched.

The idea behind here is to be able to implement also Quarkus, Micronaut and Ktor implementations of the same app, replacing only the framework.